+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 2.3.3 (Unreleased)
 ---------------------
 - Enh #336: Display page visibility icon
+- Enh #333: Added config option `Hide Navigation Entry`
 
 2.3.2 (June 25, 2024)
 ---------------------